feat: runtime demo v1.3 — sound overhaul, title screen, boot intro
Sound system:
- Eagle fanfare (track 4) now layers FM_VICTORY_JINGLE + SOUND_BED_TITLE_PLUCK over SSG
- Warrior battle (track 3) uses FM_DUEL_SUSPENSE + SSG_WARRIOR_RIFF, T160
- Game loop starts on SAMURAI_GAME_LOOP; first attack switches to WARRIOR_BATTLE via MUSIC() action cmd
- init_scene() runs before soundPlayGameLoop() — correct startup order
- ADPCMA boosted to 0x3F after game loop mix for crisp SFX
- ADPCMB volume fixed to 0xB8 in intro and title (was 0xF0, causing distortion)

Boot intro (showEagleIntro):
- Fanfare playMusic restored; debug FM call removed
- Voice cues GET_READY / ATTACK restored (once per word, not per-letter)
- Typewriter: 70ms/60ms per character — readable cadence
- Title gong SFX on logo reveal; proper fade + setBACKDROP(BLACK)

Title screen (showTitleScreen):
- INSERT COIN blink loop restored (30-frame on/off, 6 seconds)
- FIX palette fixed: pal 0=YELLOW, 1=WHITE, 2=CYAN text (was BLACK/invisible)
- soundFadeOutSpeed(4) on exit instead of abrupt cut

Artbox: 83 warrior/eagle/fx sprite source PNGs added; genscreens compact mode
SDK: ng_defs utility macros, ng_actions docs, ng_chars helpers, sound_ids fanfare ID

+++ sdk/ng_actions.h
@@ -5,6 +5,31 @@
 
 struct NGCharacter;
 
+/*
+ * NGActionCmd — one step in an animation script.
+ *
+ * Scripts are const arrays terminated by END() or LOOP().
+ * The engine processes one command per frame pass; FRAME() commands consume
+ * multiple frames (the `time` argument is the hold duration in ticks).
+ *
+ * Quick reference:
+ *   FRAME(tile, ticks)   — display tileBase `tile` for `ticks` frames
+ *   PALETTE(bank)        — switch to palette bank `bank`
+ *   FLIP(hflip, vflip)   — set mirror flags (0 or 1)
+ *   SCALE(sx, sy)        — set scale (0xFF = full size)
+ *   MOVE(dx, dy)         — instant position offset
+ *   SPEED(vx, vy)        — set fixed-point velocity (pixels/frame)
+ *   WAIT(ticks)          — pause script for `ticks` frames (no tile change)
+ *   SFX(id)              — fire ADPCM-A sample via the registered SFX hook
+ *   MUSIC(id)            — start music track via the registered music hook
+ *   FX(kind, dx, dy)     — call the FX hook at (c->x+dx, c->y+dy)
+ *   HITBOX(x,y,w,h)      — enable a hit-stun box relative to the character
+ *   CLEAR_HITBOX()       — remove the hit-stun box
+ *   EVENT(id,a,b)        — fire a game event with payload (a, b)
+ *   GOTO(action_id)      — jump to another registered action script
+ *   LOOP()               — restart the current script from the beginning
+ *   END()                — stop the script; character holds its last frame
+ */
 typedef struct {
     uint8_t cmd;
     int16_t a;
@@ -32,22 +57,22 @@ enum {
     ACT_GOTO
 };
 
-#define FRAME(tile,time)      { ACT_FRAME, tile, time, 0, 0 }
-#define WAIT(time)            { ACT_WAIT, time, 0, 0, 0 }
-#define MOVE(dx,dy)           { ACT_MOVE, dx, dy, 0, 0 }
-#define SPEED(vx,vy)          { ACT_SPEED, vx, vy, 0, 0 }
-#define SCALE(sx,sy)          { ACT_SCALE, sx, sy, 0, 0 }
-#define PALETTE(pal)          { ACT_PALETTE, pal, 0, 0, 0 }
-#define FLIP(x,y)             { ACT_FLIP, x, y, 0, 0 }
-#define SFX(id)               { ACT_SFX, id, 0, 0, 0 }
-#define MUSIC(id)             { ACT_MUSIC, id, 0, 0, 0 }
-#define FX(kind,x,y)          { ACT_FX, kind, x, y, 0 }
-#define HITBOX(x,y,w,h)       { ACT_HITBOX, x, y, w, h }
+#define FRAME(tile,time)      { ACT_FRAME,       (int16_t)(tile), (int16_t)(time), 0, 0 }
+#define WAIT(time)            { ACT_WAIT,         (int16_t)(time), 0, 0, 0 }
+#define MOVE(dx,dy)           { ACT_MOVE,         (int16_t)(dx),   (int16_t)(dy), 0, 0 }
+#define SPEED(vx,vy)          { ACT_SPEED,        (int16_t)(vx),   (int16_t)(vy), 0, 0 }
+#define SCALE(sx,sy)          { ACT_SCALE,        (int16_t)(sx),   (int16_t)(sy), 0, 0 }
+#define PALETTE(pal)          { ACT_PALETTE,      (int16_t)(pal),  0, 0, 0 }
+#define FLIP(x,y)             { ACT_FLIP,         (int16_t)(x),    (int16_t)(y), 0, 0 }
+#define SFX(id)               { ACT_SFX,          (int16_t)(id),   0, 0, 0 }
+#define MUSIC(id)             { ACT_MUSIC,        (int16_t)(id),   0, 0, 0 }
+#define FX(kind,x,y)          { ACT_FX,           (int16_t)(kind), (int16_t)(x), (int16_t)(y), 0 }
+#define HITBOX(x,y,w,h)       { ACT_HITBOX,       (int16_t)(x),    (int16_t)(y), (int16_t)(w), (int16_t)(h) }
 #define CLEAR_HITBOX()        { ACT_CLEAR_HITBOX, 0, 0, 0, 0 }
-#define EVENT(id,a,b)         { ACT_EVENT, id, a, b, 0 }
-#define GOTO(action)          { ACT_GOTO, action, 0, 0, 0 }
-#define LOOP()                { ACT_LOOP, 0, 0, 0, 0 }
-#define END()                 { ACT_END, 0, 0, 0, 0 }
+#define EVENT(id,a,b)         { ACT_EVENT,        (int16_t)(id),   (int16_t)(a), (int16_t)(b), 0 }
+#define GOTO(action)          { ACT_GOTO,         (int16_t)(action), 0, 0, 0 }
+#define LOOP()                { ACT_LOOP,         0, 0, 0, 0 }
+#define END()                 { ACT_END,          0, 0, 0, 0 }
 
 typedef void (*NGActionSfxHook)(uint16_t id);
 typedef void (*NGActionMusicHook)(uint16_t id);
